+++ src/main/java/com/invadermonky/magicultureintegrations/integrations/astralsorcery/events/ASCommonEventHandler.java
@@ -13,7 +13,7 @@ import net.minecraftforge.fluids.FluidRegistry;
 import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
 
-public class ASEventSubscriber {
+public class ASCommonEventHandler {
     @SubscribeEvent
     public void onRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
         World world = event.getWorld();