+++ src/main/java/com/invadermonky/magicultureintegrations/integrations/thaumcraft/InitThaumcraft.java
@@ -1,7 +1,10 @@
 package com.invadermonky.magicultureintegrations.integrations.thaumcraft;
 
 import com.invadermonky.magicultureintegrations.api.mods.IntegrationModule;
+import com.invadermonky.magicultureintegrations.config.MIConfigAdditions;
 import com.invadermonky.magicultureintegrations.config.MIConfigIntegrations;
+import com.invadermonky.magicultureintegrations.integrations.thaumcraft.infusion.InfusionEnchantRevealing;
+import com.invadermonky.magicultureintegrations.registry.RegistrarMI;
 import thaumcraft.common.tiles.devices.TileBellows;
 
 public class InitThaumcraft extends IntegrationModule {
@@ -11,6 +14,13 @@ public class InitThaumcraft extends IntegrationModule {
 
     @Override
     public void buildModIntegrations() {
+    }
+
+    @Override
+    public void preInit() {
+        if (MIConfigAdditions.thaumcraft.enableRevealingInfusion) {
+            RegistrarMI.registerAddition(new InfusionEnchantRevealing());
+        }
     }
 
     @Override